Исправлены названия view
