+++ src/presenter/films-presenter.js
@@ -1,10 +1,10 @@
 import {render} from '../render.js';
-import CardView from '../view/card.js';
-import FilmListContainerView from '../view/film-list-container.js';
-import FilmListView from '../view/film-list.js';
-import FilmSectionView from '../view/film-section.js';
-import HiddenTitleView from '../view/hidden-title-movies.js';
-import ShowMoreButtonView from '../view/show-more-btn.js';
+import CardView from '../view/card-view.js';
+import FilmListContainerView from '../view/film-list-container-view.js';
+import FilmListView from '../view/film-list-view.js';
+import FilmSectionView from '../view/film-section-view.js';
+import HiddenTitleView from '../view/hidden-title-view.js';
+import ShowMoreButtonView from '../view/show-more-button-view.js';
 
 export default class FilmsPresenter {
   filmSection = new FilmSectionView();
@@ -18,7 +18,7 @@ export default class FilmsPresenter {
   init() {
     render(this.filmSection, this.filmsContainer);
     render(this.filmList, this.filmSection.getElement());
-    render(new HiddenTitleView(), this.filmSection.getElement());
+    render(new HiddenTitleView(), this.filmList.getElement());
     render(this.filmListContainer, this.filmList.getElement());
 
     for (let i = 0; i < 5; i++) {
