+++ src/view/card-view.js
@@ -1,0 +1,44 @@
+import {createElement} from '../render.js';
+
+function createCardTemplate() {
+  return (
+    `<article class="film-card">
+      <a class="film-card__link">
+        <h3 class="film-card__title">The Dance of Life</h3>
+        <p class="film-card__rating">8.3</p>
+        <p class="film-card__info">
+          <span class="film-card__year">1929</span>
+          <span class="film-card__duration">1h 55m</span>
+          <span class="film-card__genre">Musical</span>
+        </p>
+        <img src="./images/posters/the-dance-of-life.jpg" alt="" class="film-card__poster">
+        <p class="film-card__description">Burlesque comic Ralph "Skid" Johnson (Skelly), and specialty dancer Bonny Lee King (Carroll), end up together on a cold, rainy night at a tr…</p>
+        <span class="film-card__comments">5 comments</span>
+      </a>
+      <div class="film-card__controls">
+        <button class="film-card__controls-item film-card__controls-item--add-to-watchlist" type="button">Add to watchlist</button>
+        <button class="film-card__controls-item film-card__controls-item--mark-as-watched" type="button">Mark as watched</button>
+        <button class="film-card__controls-item film-card__controls-item--favorite" type="button">Mark as favorite</button>
+      </div>
+    </article>
+    `
+  );
+}
+
+export default class CardView {
+  getTemplate() {
+    return createCardTemplate();
+  }
+
+  getElement() {
+    if (!this.element) {
+      this.element = createElement(this.getTemplate());
+    }
+
+    return this.element;
+  }
+
+  removeElement() {
+    this.element = null;
+  }
+}
