+++ src/view/film-section-view.js
