+++ src/view/hidden-title-view.js
