+++ src/view/rank-user-view.js
@@ -2,10 +2,11 @@ import {createElement} from '../render.js';
 
 function createRankUserTemplate() {
   return (
-    `  <section class="header__profile profile">
-    <p class="profile__rating">Movie Buff</p>
-    <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>`
+    `<section class="header__profile profile">
+      <p class="profile__rating">Movie Buff</p>
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>
+    `
   );
 }
