+++ src/view/show-more-button-view.js
