+++ src/view/statistic-view.js
